requirements.txt
instaloader
pydrive
requests[socks]
PySocks 